refactor (webapi): move methods to test package
